+++ frontend/src/components/landing/LandingServices.jsx
@@ -1,0 +1,19 @@
+export default function LandingServices() {
+    return (
+        <div className="container">
+            <h1>This site is a work in progress, but plans to add the following:</h1>
+            <ul className="lead text-start mt-5">
+                <li>Personal and Professional Social Media Connections</li>
+                <li>Local Event Listings and RSVPs</li>
+                <li>Community Forums and Discussion Boards</li>
+                <li>Personal and Business Shopping</li>
+                <li>Local News and Updates</li>
+                <li>Service Provider Listings and Reviews</li>
+                <li>Volunteer Opportunities and Community Projects</li>
+                <li>Local Weather</li>
+                <li>And Much More!</li>
+            </ul>
+            <p className="lead mt-4">Stay tuned for updates as we work towards bringing these features to life and enhancing the way local communities connect and engage!</p>
+        </div>
+    );
+}